requirements.txt
amara3.xml

rdflib
versa
diskcache

#Optional bits
#CacheControl[filecache]
#rdflib-jsonld
